Move test runner code into modules
